pyproject.toml
[tool.poetry]
name = "StackWar"
version = "1.0.0"
description = ""
authors = ["Peter Hodge"]
license = "MIT"
packages = [
    { include = "stackwar" },
]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.2"
requests = "^2.24.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
flake8 = "^3.8.3"
pynvim = "^0.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
